Ver 2.1 - Introduction of tabs to the output display and the "Function Calculator" layout
