+++ MathsVisualisationTool/resources/App.xaml
@@ -4,18 +4,18 @@
              xmlns:local="clr-namespace:MathsVisualisationTool"
              StartupUri="../views/MainWindow.xaml">
     <Application.Resources>
-        <!-- Style sheet -->
+         <!--Style sheet--> 
         
-        <!-- Dock Panel style --><!--
+         <!--Dock Panel style--> 
         <Style x:Key="dockPanelStyle">
             <Setter Property="DockPanel.Background" Value="#170181"/>
         </Style>
         
-        --><!-- Top Menu style --><!--
+         <!--Top Menu style--> 
         <Style x:Key="menuStyle">
             <Setter Property="DockPanel.Dock" Value="Top"/>
             <Setter Property="Menu.FontSize" Value="14"/>
-        </Style>-->
+        </Style>
 
     </Application.Resources>
 </Application>
